[core] Simplify FileStoreCommitImpl to extract some classes
